refactor(other): replace Float32 flags and ParseFloat with integer types

+++ cmd/search/multi.go
@@ -18,12 +18,12 @@ var multiCmd = &cobra.Command{
 		apiClient, ctx, isVerbose := newAPIClient()
 
 		query, _ := cmd.Flags().GetString("query")
-		page, _ := cmd.Flags().GetFloat32("page")
+		page, _ := cmd.Flags().GetInt("page")
 		language, _ := cmd.Flags().GetString("language")
 
 		req := apiClient.SearchAPI.SearchGet(ctx).Query(query)
 		if cmd.Flags().Changed("page") {
-			req = req.Page(page)
+			req = req.Page(float32(page))
 		}
 		if cmd.Flags().Changed("language") {
 			req = req.Language(language)
@@ -37,7 +37,7 @@ var multiCmd = &cobra.Command{
 func init() {
 	multiCmd.Flags().StringP("query", "q", "", "Search query")
 	multiCmd.MarkFlagRequired("query")
-	multiCmd.Flags().Float32("page", 1, "Page number")
+	multiCmd.Flags().Int("page", 1, "Page number")
 	multiCmd.Flags().String("language", "en", "Language code")
 	Cmd.AddCommand(multiCmd)
 }
